+++ include/isaac/jit/generation/elementwise_2d.h
@@ -36,12 +36,12 @@ private:
   int is_invalid_impl(driver::Device const &, expression_tree const  &) const;
   std::string generate_impl(std::string const & suffix, expression_tree const  & expressions, driver::Device const & device, symbolic::symbols_table const & mapping) const;
 public:
-  elementwise_2d(unsigned int simd, unsigned int ls0, unsigned int ls1,  unsigned int ng0, unsigned int ng1, fetching_policy_type fetch);
+  elementwise_2d(size_t simd, size_t ls0, size_t ls1,  size_t ng0, size_t ng1, fetching_policy_type fetch);
   std::vector<int_t> input_sizes(expression_tree const  & expressions) const;
   void enqueue(driver::CommandQueue & queue, driver::Program const & program, std::string const & suffix, expression_tree const & tree, runtime::environment const & opt);
 private:
-  unsigned int num_groups_0;
-  unsigned int num_groups_1;
+  size_t num_groups_0;
+  size_t num_groups_1;
   fetching_policy_type fetching_policy;
 };
